+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Internal constant `HexaPDF::Content::ColorSpace::CSS_COLOR_NAMES` changed to
   [HexaPDF::Content::ColorSpace::COLOR_NAMES]
+* Constructor of [HexaPDF::Layout::PageStyle] to allow setting `next_style`
+  attribute
 
 
 ## 0.32.2 - 2023-05-06